reorganize detector subpackage into instrument and optics
