+++ mocksipipeline/instrument/optics/data/hetgD1996-11-01greffpr001N0007.fits
