+++ mocksipipeline/instrument/optics/filter.py
