+++ setup.cfg
@@ -45,7 +45,7 @@ docs =
     sphinx-automodapi
 
 [options.package_data]
-mocksipipeline.detector = data/hetgD1996-11-01greffpr001N0007.fits
+mocksipipeline.instrument.optics = data/hetgD1996-11-01greffpr001N0007.fits
 mocksipipeline.spectral = data/*.asdf
 
 [tool:pytest]
